reorganize passes directory
