+++ MANIFEST.in
@@ -1,6 +1,6 @@
 include qiskit/qasm/libs/*.inc
 include qiskit/schemas/*json
 include qiskit/VERSION.txt
-include qiskit/transpiler/passes/mapping/cython/stochastic_swap/*pyx
-include qiskit/transpiler/passes/mapping/cython/stochastic_swap/*pxd
+include qiskit/transpiler/passes/routing/cython/stochastic_swap/*pyx
+include qiskit/transpiler/passes/routing/cython/stochastic_swap/*pxd
 recursive-include qiskit/test/mock/backends *json
