+++ qiskit/circuit/quantumcircuit.py
@@ -607,7 +607,7 @@ class QuantumCircuit:
         Returns:
             QuantumCircuit: a circuit one level decomposed
         """
-        from qiskit.transpiler.passes.decompose import Decompose
+        from qiskit.transpiler.passes.basis.decompose import Decompose
         from qiskit.converters.circuit_to_dag import circuit_to_dag
         from qiskit.converters.dag_to_circuit import dag_to_circuit
         pass_ = Decompose()
