+++ qiskit/transpiler/passes/__init__.py
@@ -19,32 +19,24 @@ Transpiler Passes (:mod:`qiskit.transpiler.passes`)
 
 .. currentmodule:: qiskit.transpiler.passes
 
-Layout Selection
-================
+Layout Selection (Placement)
+============================
 
 .. autosummary::
    :toctree: ../stubs/
 
+   SetLayout
    TrivialLayout
    DenseLayout
    NoiseAdaptiveLayout
+   CSPLayout
    ApplyLayout
-   SetLayout
+   Layout2qDistance
    EnlargeWithAncilla
    FullAncillaAllocation
 
-Unrolling
-=========
-
-.. autosummary::
-   :toctree: ../stubs/
-
-   Unroller
-   Unroll3qOrMore
-   Decompose
-
-Swap Mapping
-============
+Routing
+=======
 
 .. autosummary::
    :toctree: ../stubs/
@@ -53,38 +45,34 @@ Swap Mapping
    LookaheadSwap
    StochasticSwap
 
-Gate Optimizations
-==================
+Basis Change
+============
+
+.. autosummary::
+   :toctree: ../stubs/
+
+   Unroller
+   Unroll3qOrMore
+   Decompose
+
+Optimizations
+=============
 
 .. autosummary::
    :toctree: ../stubs/
 
    Optimize1qGates
+   Collect2qBlocks
+   ConsolidateBlocks
    CXCancellation
    CommutationAnalysis
    CommutativeCancellation
+   RemoveDiagonalGatesBeforeMeasure
+   RemoveResetInZeroState
+   CrosstalkAdaptiveSchedule
 
-2Q Block Optimization
-=====================
-
-.. autosummary::
-   :toctree: ../stubs/
-
-   Collect2qBlocks
-   ConsolidateBlocks
-
-Topology Validation
-===================
-
-.. autosummary::
-   :toctree: ../stubs/
-
-   CheckMap
-   CheckCXDirection
-   CXDirection
-
-DAG Properties
-==============
+Circuit Analysis
+================
 
 .. autosummary::
    :toctree: ../stubs/
@@ -95,7 +83,6 @@ DAG Properties
    CountOps
    CountOpsLongestPath
    NumTensorFactors
-   DAGFixedPoint
    DAGLongestPath
 
 Additional Passes
@@ -104,52 +91,65 @@ Additional Passes
 .. autosummary::
    :toctree: ../stubs/
 
+   CheckMap
+   CheckCXDirection
+   CXDirection
    MergeAdjacentBarriers
    BarrierBeforeFinalMeasurements
-   RemoveDiagonalGatesBeforeMeasure
    RemoveFinalMeasurements
-   RemoveResetInZeroState
-   CrosstalkAdaptiveSchedule
+   DAGFixedPoint
+   FixedPoint
 """
 
-from .unroller import Unroller
-from .cx_cancellation import CXCancellation
-from .fixed_point import FixedPoint
-from .resource_estimation import ResourceEstimation
-from .depth import Depth
-from .size import Size
-from .width import Width
-from .count_ops import CountOps
-from .count_ops_longest_path import CountOpsLongestPath
-from .num_tensor_factors import NumTensorFactors
-from .dag_fixed_point import DAGFixedPoint
-from .dag_longest_path import DAGLongestPath
-from .merge_adjacent_barriers import MergeAdjacentBarriers
-from .optimize_1q_gates import Optimize1qGates
-from .decompose import Decompose
-from .unroll_3q_or_more import Unroll3qOrMore
-from .commutation_analysis import CommutationAnalysis
-from .optimize_swap_before_measure import OptimizeSwapBeforeMeasure
-from .commutative_cancellation import CommutativeCancellation
-from .remove_reset_in_zero_state import RemoveResetInZeroState
-from .collect_2q_blocks import Collect2qBlocks
-from .consolidate_blocks import ConsolidateBlocks
-from .barrier_before_final_measurements import BarrierBeforeFinalMeasurements
-from .mapping.full_ancilla_allocation import FullAncillaAllocation
-from .mapping.enlarge_with_ancilla import EnlargeWithAncilla
-from .mapping.apply_layout import ApplyLayout
-from .mapping.check_map import CheckMap
-from .mapping.check_cx_direction import CheckCXDirection
-from .mapping.cx_direction import CXDirection
-from .mapping.trivial_layout import TrivialLayout
-from .mapping.csp_layout import CSPLayout
-from .mapping.set_layout import SetLayout
-from .mapping.dense_layout import DenseLayout
-from .mapping.noise_adaptive_layout import NoiseAdaptiveLayout
-from .mapping.basic_swap import BasicSwap
-from .mapping.layout_2q_distance import Layout2qDistance
-from .mapping.lookahead_swap import LookaheadSwap
-from .remove_diagonal_gates_before_measure import RemoveDiagonalGatesBeforeMeasure
-from .mapping.stochastic_swap import StochasticSwap
-from .remove_final_measurements import RemoveFinalMeasurements
-from .crosstalk_adaptive_schedule import CrosstalkAdaptiveSchedule
+# layout selection (placement)
+from .layout import SetLayout
+from .layout import TrivialLayout
+from .layout import DenseLayout
+from .layout import NoiseAdaptiveLayout
+from .layout import CSPLayout
+from .layout import ApplyLayout
+from .layout import Layout2qDistance
+from .layout import EnlargeWithAncilla
+from .layout import FullAncillaAllocation
+
+# routing
+from .routing import BasicSwap
+from .routing import LookaheadSwap
+from .routing import StochasticSwap
+
+# basis change
+from .basis import Decompose
+from .basis import Unroller
+from .basis import Unroll3qOrMore
+
+# optimization
+from .optimization import Optimize1qGates
+from .optimization import Collect2qBlocks
+from .optimization import ConsolidateBlocks
+from .optimization import CommutationAnalysis
+from .optimization import CommutativeCancellation
+from .optimization import CXCancellation
+from .optimization import OptimizeSwapBeforeMeasure
+from .optimization import RemoveResetInZeroState
+from .optimization import RemoveDiagonalGatesBeforeMeasure
+from .optimization import CrosstalkAdaptiveSchedule
+
+# circuit analysis
+from .analysis import ResourceEstimation
+from .analysis import Depth
+from .analysis import Size
+from .analysis import Width
+from .analysis import CountOps
+from .analysis import CountOpsLongestPath
+from .analysis import NumTensorFactors
+from .analysis import DAGLongestPath
+
+# additional utility passes
+from .utils import CheckMap
+from .utils import CheckCXDirection
+from .utils import CXDirection
+from .utils import BarrierBeforeFinalMeasurements
+from .utils import RemoveFinalMeasurements
+from .utils import MergeAdjacentBarriers
+from .utils import DAGFixedPoint
+from .utils import FixedPoint
