+++ qiskit/transpiler/passes/analysis/count_ops.py
