+++ qiskit/transpiler/passes/analysis/depth.py
