+++ qiskit/transpiler/passes/analysis/num_qubits.py
