+++ qiskit/transpiler/passes/analysis/num_tensor_factors.py
