+++ qiskit/transpiler/passes/analysis/size.py
