+++ qiskit/transpiler/passes/analysis/width.py
