+++ qiskit/transpiler/passes/layout/__init__.py
@@ -12,4 +12,14 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-"""Module containing transpiler mapping passes."""
+"""Module containing transpiler layout passes."""
+
+from .set_layout import SetLayout
+from .trivial_layout import TrivialLayout
+from .dense_layout import DenseLayout
+from .noise_adaptive_layout import NoiseAdaptiveLayout
+from .csp_layout import CSPLayout
+from .apply_layout import ApplyLayout
+from .layout_2q_distance import Layout2qDistance
+from .enlarge_with_ancilla import EnlargeWithAncilla
+from .full_ancilla_allocation import FullAncillaAllocation
