+++ qiskit/transpiler/passes/optimization/collect_2q_blocks.py
