+++ qiskit/transpiler/passes/optimization/commutation_analysis.py
