+++ qiskit/transpiler/passes/optimization/consolidate_blocks.py
