+++ qiskit/transpiler/passes/optimization/cx_cancellation.py
