+++ qiskit/transpiler/passes/optimization/optimize_1q_gates.py
