+++ qiskit/transpiler/passes/optimization/remove_diagonal_gates_before_measure.py
