+++ qiskit/transpiler/passes/optimization/remove_reset_in_zero_state.py
