+++ qiskit/transpiler/passes/routing/cython/stochastic_swap/swap_trial.pyx
