+++ qiskit/transpiler/passes/routing/cython/stochastic_swap/utils.pyx
