+++ qiskit/transpiler/passes/utils/fixed_point.py
