+++ qiskit/transpiler/passes/utils/merge_adjacent_barriers.py
